# Conflicts:
#	Client/client_controller.cpp
#	Client/client_view.cpp
#	Event/event_type.h
#	Model/game_data_model.h
#	Server/server_controller.cpp
#	constants.h

+++ Model/game_data_model.cpp
@@ -27,8 +27,8 @@ int GameDataModel::GetPlayersCount() const {
 }
 
 void GameDataModel::AddPlayer(GameObjectId player_id,
-                              int x,
-                              int y,
+                              float x,
+                              float y,
                               float view_angle) {
   if (players_.find(player_id) != players_.end()) {
     throw std::runtime_error("[MODEL] This player_id already exists");
